+++ stock_monitor/utils/__init__.py
@@ -2,7 +2,7 @@
 
 from .helpers import get_stock_emoji, is_equal, resource_path
 from .logger import app_logger
-from .stock_utils import StockCodeProcessor, extract_stocks_from_list
+from .stock_utils import StockCodeProcessor
 
 __all__ = [
     "resource_path",
@@ -10,5 +10,4 @@ __all__ = [
     "is_equal",
     "app_logger",
     "StockCodeProcessor",
-    "extract_stocks_from_list",
 ]